+++ lib/iris/ugrid/__init__.py
@@ -21,7 +21,7 @@ Based on CF UGRID Conventions (v1.0), https://ugrid-conventions.github.io/ugrid-
 """
 
 from ..config import get_logger
-from .load import PARSE_UGRID_ON_LOAD, load_mesh, load_meshes
+from .load import load_mesh, load_meshes
 from .mesh import Connectivity, MeshCoord, MeshXY
 from .save import save_mesh
 from .utils import recombine_submeshes
@@ -30,7 +30,6 @@ __all__ = [
     "Connectivity",
     "MeshCoord",
     "MeshXY",
-    "PARSE_UGRID_ON_LOAD",
     "load_mesh",
     "load_meshes",
     "recombine_submeshes",
